Implements pick preview handling with updateCard controller Abstracts truncating picks in card to own function fo use in create and update card controllers
